services/screen-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
mss==9.0.1
Pillow==10.2.0
python-multipart==0.0.6
websockets==12.0
